improve building of package
