+++ package.json
@@ -15,12 +15,14 @@
 		"dev": "turbo dev",
 		"test": "turbo test",
 		"format": "prettier --write .",
-		"sherif": "sherif"
+		"sherif": "sherif",
+		"publint": "publint"
 	},
 	"devDependencies": {
 		"prettier": "^3.0.3",
 		"prettier-plugin-astro": "^0.12.0",
 		"prettier-plugin-tailwindcss": "^0.5.5",
+		"publint": "^0.2.3",
 		"sherif": "^0.3.1",
 		"turbo": "^1.10.15",
 		"typescript": "^5.2.2"
