+++ packages/palette/tsconfig.json
@@ -5,9 +5,6 @@
 		"declaration": true,
 		"emitDeclarationOnly": true,
 		"noEmit": false,
-		"outDir": "dist",
-		"types": [
-			"vitest/importMeta"
-		]
+		"outDir": "dist"
 	}
 }
