--- packages/palette/vitest.config.ts
@@ -1,7 +1,0 @@
-import { defineConfig } from "vitest/config";
-
-export default defineConfig({
-	test: {
-		includeSource: ["src/**/*.ts"],
-	},
-});
